+++ app/api/gpt-logs/route.ts
@@ -2,6 +2,7 @@ import { db } from "@/lib/db";
 
 export const dynamic = "force-dynamic";
 
+// todo: convert to trpc
 export async function POST(req: Request) {
   const { data } = await req.json();
   // Create a new requestResponse in the database, but don't await it
