+++ server/api/routers/user.ts
@@ -99,4 +99,47 @@ export const userRouter = createTRPCRouter({
         },
       });
     }),
+  getTopUsersByUpcomingEvents: publicProcedure
+    .input(z.object({ limit: z.number() }))
+    .query(async ({ ctx, input }) => {
+      const excludeUsers = ["user_2X3xAXHdaKKG8RLZqm72wb119Yj"];
+      const currentDate = new Date();
+
+      const leaderboardUsers = await ctx.db.user.findMany({
+        where: {
+          id: {
+            notIn: excludeUsers,
+          },
+        },
+        select: {
+          id: true,
+          username: true,
+          displayName: true,
+          imageUrl: true,
+          _count: {
+            select: {
+              events: {
+                where: {
+                  startDateTime: {
+                    gt: currentDate,
+                  },
+                },
+              },
+            },
+          },
+        },
+        orderBy: {
+          events: {
+            _count: "desc",
+          },
+        },
+      });
+
+      const sortedLeaderboardUsers = leaderboardUsers
+        .sort((a, b) => b._count.events - a._count.events)
+        .slice(0, input.limit)
+        .filter((user) => user._count.events > 0);
+
+      return sortedLeaderboardUsers;
+    }),
 });
